backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
pyjwt==2.8.0
python-dotenv==1.0.0 
 bcrypt-==4.3.0